refactor: ProfitCalculator 패키지 변경
