+++ src/test/java/blackjack/domain/result/ProfitCalculatorTest.java
@@ -1,6 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
-import blackjack.domain.result.GameResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
